# Cherry-pick #22974 to 7.x: Improve event normalization performance